chore(VLE): Clean up some notification functions
#454
