+++ src/assets/wise5/services/notificationService.ts
@@ -1,12 +1,12 @@
 import * as angular from 'angular';
 import { Injectable } from '@angular/core';
-import { UpgradeModule } from '@angular/upgrade/static';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { ConfigService } from './configService';
 import { ProjectService } from './projectService';
 import { UtilService } from './utilService';
 import { Notification } from '../../../app/domain/notification';
 import { Observable, Subject } from 'rxjs';
+import { AnnotationService } from './annotationService';
 
 @Injectable()
 export class NotificationService {
@@ -23,7 +23,7 @@ export class NotificationService {
   public viewCurrentAmbientNotification$: Observable<any> = this.viewCurrentAmbientNotificationSource.asObservable();
 
   constructor(
-    private upgrade: UpgradeModule,
+    private annotationService: AnnotationService,
     private http: HttpClient,
     private ConfigService: ConfigService,
     private ProjectService: ProjectService,
@@ -102,6 +102,68 @@ export class NotificationService {
         });
         return this.notifications;
       });
+  }
+
+  getNewNotifications(): any[] {
+    let newNotificationAggregates = [];
+    for (const notification of this.notifications) {
+      if (notification.timeDismissed == null) {
+        let notificationNodeId = notification.nodeId;
+        let notificationType = notification.type;
+        let newNotificationForNodeIdAndTypeExists = false;
+        for (const newNotificationAggregate of newNotificationAggregates) {
+          if (
+            newNotificationAggregate.nodeId == notificationNodeId &&
+            newNotificationAggregate.type == notificationType
+          ) {
+            newNotificationForNodeIdAndTypeExists = true;
+            newNotificationAggregate.notifications.push(notification);
+            if (notification.timeGenerated > newNotificationAggregate.latestNotificationTimestamp) {
+              newNotificationAggregate.latestNotificationTimestamp = notification.timeGenerated;
+            }
+          }
+        }
+        let notebookItemId = null; // if this notification was created because teacher commented on a notebook report.
+        if (!newNotificationForNodeIdAndTypeExists) {
+          let message = '';
+          if (notificationType === 'DiscussionReply') {
+            message = $localize`You have new replies to your discussion post!`;
+          } else if (notificationType === 'teacherToStudent') {
+            message = $localize`You have new feedback from your teacher!`;
+            if (notification.data != null) {
+              if (typeof notification.data === 'string') {
+                notification.data = JSON.parse(notification.data);
+              }
+
+              const annotationId = (notification.data as any).annotationId;
+              if (annotationId != null) {
+                let annotation = this.annotationService.getAnnotationById(annotationId);
+                if (annotation != null && annotation.notebookItemId != null) {
+                  notebookItemId = annotation.notebookItemId;
+                }
+              }
+            }
+          } else if (notificationType === 'CRaterResult') {
+            message = $localize`You have new feedback!`;
+          }
+          const newNotificationAggregate = {
+            latestNotificationTimestamp: notification.timeGenerated,
+            message: message,
+            nodeId: notificationNodeId,
+            notebookItemId: notebookItemId,
+            notifications: [notification],
+            type: notificationType
+          };
+          newNotificationAggregates.push(newNotificationAggregate);
+        }
+      }
+    }
+
+    // sort the aggregates by latestNotificationTimestamp, latest -> oldest
+    newNotificationAggregates.sort((n1, n2) => {
+      return n2.latestNotificationTimestamp - n1.latestNotificationTimestamp;
+    });
+    return newNotificationAggregates;
   }
 
   setNotificationNodePositionAndTitle(notification: Notification) {
